[contrib] skip the first conflict when rebasing the electrum history
Summary:
The first time this script was launched, it worked properly, but due to the way we integrated electrum into the monorepo (a copy of all files) the initial commit conflicts with itself on subsequent runs. This conflict can be skipped, as we are already confident that the github version of that commit was properly rebased the first time.

Note that `git rebase --skip` will exit with error code 1 if there is another conflict after the one that is skipped, so we will see the script fail if there is ever another conflict to handle (which should not happen if this remains the only source for new github commits)

Test Plan:
`contrib/extract-electrum.sh && cd electrumabc-mirror && git status`
Check that everything is up to date and that there is not another conflict to be solved.

Reviewers: #bitcoin_abc, Fabien, bytesofman

Reviewed By: #bitcoin_abc, Fabien, bytesofman

Subscribers: bytesofman

Differential Revision: https://reviews.bitcoinabc.org/D14384