test: Avoid intermittent issues due to async events in validationinterface_tests
Summary:
Currently the tests have many issues:

- They setup the genesis block, even though it is not needed
- They queue an async UpdatedBlockTip even, which causes intermittent issues: https://github.com/bitcoin/bitcoin/issues/28146#issuecomment-1649931679

This is a backport of core#28150

Test Plan:
I ran `./src/test/test_bitcoin -t validationinterface_tests/unregister_all_during_call -- -printtoconsole=1` in a loop but was not able to reproduce the intermittent failure locallly.
We can monitor the CI and see if the test keeps failing.

Reviewers: #bitcoin_abc, Fabien

Reviewed By: #bitcoin_abc, Fabien

Differential Revision: https://reviews.bitcoinabc.org/D14389