[explorer] Properly handle the search bar
Summary:
The current search feature is is split between the rust server search function and some hacky js which is mostly forwarding the request, plus:
 - Allow for searching by block height
 - Allow for prefixless addresses

This diff fixes both of the missing search features while removing the hacky js code. This makes it possible to change the prefix and keep the explorer search feature working.

Test Plan: Run the explorer and search for block height, block hash, txid, address with and without prefix (including etoken: addresses).

Reviewers: #bitcoin_abc, tobias_ruck

Reviewed By: #bitcoin_abc, tobias_ruck

Subscribers: tobias_ruck

Differential Revision: https://reviews.bitcoinabc.org/D14450