[avalanche] Improve the staking reward proof age selection
Summary:
The current algo filters the proof that are too recent with respect to the actual time the algo is run. This could lead to some jitter in the proof set as the reference point is not common across the network.

This diff improves on that by using the previous block time as a time reference. There is still an edge case that should be addressed: in the event the block time is in the future, we fall back to the current time.

Test Plan:
  ninja all check-all

Reviewers: #bitcoin_abc, PiRK

Reviewed By: #bitcoin_abc, PiRK

Differential Revision: https://reviews.bitcoinabc.org/D14465