[electrum] use TxInput (almost) everywhere in transaction.py
Summary:
This makes the Transaction class use TxInput everywhere internally, except for a few remaining methods that work with `fetched_inputs` which still need to be converted.

Input setters and their callsites are also converted:
 - `Transaction.add_inputs`
 - `Transaction.set_inputs`
 - `Transaction.update_input`
 - `Transaction.from_io`

As of this diff, `sweep` uses TxInput objects to estimate the tx size, so the test needs to be adjusted to have properly formed scriptsigs that can be parsed by TxInput.

Depends on D14510

Test Plan:
`python test_runner.py`

Send and receive transaction with the application.
Send transactions with hardware wallet.

Reviewers: #bitcoin_abc, Fabien

Reviewed By: #bitcoin_abc, Fabien

Subscribers: Fabien

Differential Revision: https://reviews.bitcoinabc.org/D14472