[alias-server] remove redundant address validation from db function
Summary:
When `getAliasInfoFromAddress` was written, it was the only function used to query at the `/address/` endpoint. So, it made sense to handle the address validation there. Now that this endpoint makes two separate database queries, the address validation should be handled before each query -- and it is, in the endpoint handling function defined in app.js

Remove address validation from `getAliasInfoFromAddress` as this function is only used by the `/address/` endpoint handler defined in app.js and this handler already performs address validation.

Test Plan: `npm test`

Reviewers: #bitcoin_abc, Fabien

Reviewed By: #bitcoin_abc, Fabien

Differential Revision: https://reviews.bitcoinabc.org/D14506