[apps and Cashtab] Implement monorepo chronik mock and deprecate repo-specific implementations
Summary:
Mocked chronik client was created initially in alias-server to support testing there. Later it was ported to other repos that needed it.

Plan is to get it into `chronik-client`. For now, though, it should at least live in only one place, so that it isn't maintained many times in the same monorepo.

Implementing.

Test Plan: `npm test` in `apps/ecash-herald`, `apps/alias-server`, and `cashtab`

Reviewers: #bitcoin_abc, Fabien

Reviewed By: #bitcoin_abc, Fabien

Subscribers: Fabien

Differential Revision: https://reviews.bitcoinabc.org/D14518