[Cashtab] Load settings page with saved wallets collapse open
Summary:
Request from cain. This probably is an improvement. Most of the time I go to settings page it is to change the wallet.

Other potential improvements: saved wallets collapse should probably appear higher on the settings screen. Better still, you should be able to change thoe active wallet from any screen, or at least the 'receive' and 'home' screens.

Test Plan: `npm test`, judgment if this is improvement, if not, if more should be done

Reviewers: #bitcoin_abc, emack

Reviewed By: #bitcoin_abc, emack

Subscribers: emack

Differential Revision: https://reviews.bitcoinabc.org/D14529