[avalanche] Don't ban for polling cooldown not elapsed
Summary:
The cooldown violation is a time dependent issue that has been observed
on honest nodes facing severe temporary network issues. This cooldown
has been introduced as an DoS prevention mechanism, and is effective at
this. Receiving a poll message is only costly if it is processed by the
node; otherwise it's equivalent to any unknown message that anybody can
already spam on the network. This diff keeps the same cooldown mechanic
but discard the message rather than banning the node, which is the same
behavior as receiving an unknown message. This also fixes the error
message having a funny typo.

Test Plan:
  ninja all check-all

Reviewers: #bitcoin_abc, bytesofman, PiRK

Reviewed By: #bitcoin_abc, bytesofman, PiRK

Subscribers: bytesofman

Differential Revision: https://reviews.bitcoinabc.org/D14556